Refactor routes
This theoretically should be two commits but since the changes are so mixed together I figured it's not that bad to commit it all at once.

The first change is to add some macros relating to routes, that automatically insert the correct signature, function attributes, try-catch blocks for HTTP::StatusException (sends the status and ends the response), msgpack exceptions (sends 400 and ends), and any other exception (sends 500 and logs the error).

The second change is to flesh out all the routes and registerers, using stubs for all the routes that aren't yet implemented. The reason for this is that registerers should hopefully never need to be modified from now on.