Use the correct constructor for `std::vector`
Rather than interpreting the `size_t` as a size, it interpreted it as an aggregate initialization of the User. This resulted in bogus entries. Instead, create the vector first, then reserve the space. This is slightly suboptimal because there may be two allocations where one would do (though such behavior is not defined by the specification). However it's (AFAICT) not possible to invoke the correct constructor.